+++ AnalysisNtuple/src/TkrHitValsTool.cxx
@@ -50,12 +50,12 @@ public:
 private:
     
     //TkrClusters Tuple Items
-    double Tkr_Cnv_Lyr_Hits;
-    double Tkr_Max_controller_hits;
-    double Tkr_Fst_Cnv_Lyr;
-    double Tkr_NCnv_Lyrs_Hit;
+    float Tkr_Cnv_Lyr_Hits;
+    float Tkr_Max_controller_hits;
+    float Tkr_Fst_Cnv_Lyr;
+    float Tkr_NCnv_Lyrs_Hit;
     
-    double Tkr_HitsPerLyr[_nLayers];
+    float Tkr_HitsPerLyr[_nLayers];
 
     ITkrQueryClustersTool* m_clusTool;
 };
